+++ test/instructions/arg.test.ts
@@ -47,4 +47,26 @@ describe("ARG", () => {
         assertRange(property.getNameRange(), 0, 4, 0, 7);
         assertRange(property.getValueRange(), 0, 8, 0, 13);
     });
+
+    it("unclosed single quote", () => {
+        let dockerfile = DockerfileParser.parse("ARG a='a\\  b");
+        let args = dockerfile.getARGs();
+        assert.equal(args.length, 1);
+        let property = args[0].getProperty();
+        assert.equal(property.getName(), "a");
+        assert.equal(property.getValue(), "'a\\  b");
+        assertRange(property.getNameRange(), 0, 4, 0, 5);
+        assertRange(property.getValueRange(), 0, 6, 0, 12);
+    });
+
+    it("unclosed double quote", () => {
+        let dockerfile = DockerfileParser.parse("ARG a=\"a\\  b");
+        let args = dockerfile.getARGs();
+        assert.equal(args.length, 1);
+        let property = args[0].getProperty();
+        assert.equal(property.getName(), "a");
+        assert.equal(property.getValue(), "\"a\\  b");
+        assertRange(property.getNameRange(), 0, 4, 0, 5);
+        assertRange(property.getValueRange(), 0, 6, 0, 12);
+    });
 });